renamed some vars, cleaned up
