+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/App';
 import reportWebVitals from './reportWebVitals';
+import App from './App';
+import './index.css';
 
 ReactDOM.render(
   <React.StrictMode>
